Implement Phase 2 fixes and enhancements across the admin panel and API. Introduced integer-only handling for coin-related fields, updated transaction status options, and improved error handling. Added new utility functions for number handling and standardized data validation. Enhanced documentation for product requirements and phase summaries.

+++ apps/admin/test/__mocks__/TransactionVerificationModal.tsx
@@ -80,7 +80,7 @@ export const TransactionVerificationModal = React.forwardRef<
   const handleApprove = () => {
     setIsLoading(true);
     onApprove?.(transaction.id, {
-      observedAmount: parseFloat(formData.observedAmount),
+      observedAmount: parseInt(formData.observedAmount),
       receiptDate: formData.receiptDate,
       verificationConfirmed: formData.verificationConfirmed,
       rejectionNote: formData.rejectionNote,
@@ -98,7 +98,7 @@ export const TransactionVerificationModal = React.forwardRef<
   const handleApproveAndPay = () => {
     setIsLoading(true);
     onApproveAndPay?.(transaction.id, {
-      observedAmount: parseFloat(formData.observedAmount),
+      observedAmount: parseInt(formData.observedAmount),
       receiptDate: formData.receiptDate,
       verificationConfirmed: formData.verificationConfirmed,
       rejectionNote: formData.rejectionNote,
